refactor: move NE sweepers to sweeper.go style
