+++ internal/resources/networkedge/device/resource_test.go
@@ -10,9 +10,10 @@ import (
 	"time"
 
 	"github.com/equinix/ne-go"
-	"github.com/equinix/terraform-provider-equinix/internal/converters"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/stretchr/testify/assert"
+
+	"github.com/equinix/terraform-provider-equinix/internal/converters"
 )
 
 func TestNetworkDevice_createFromResourceData(t *testing.T) {
