+++ internal/resources/networkedge/ssh_user/resource_acc_test.go
@@ -1,14 +1,11 @@
 package ssh_user_test
 
 import (
-	"context"
 	"fmt"
-	"log"
 
 	"github.com/equinix/terraform-provider-equinix/internal/acceptance"
 	"github.com/equinix/terraform-provider-equinix/internal/comparisons"
 	"github.com/equinix/terraform-provider-equinix/internal/config"
-	"github.com/equinix/terraform-provider-equinix/internal/sweep"
 
 	"github.com/equinix/ne-go"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
@@ -17,13 +14,6 @@ import (
 
 // SSH User acc tests are in Device acc test tests
 // reason: SSH User requires device to be provisioned and that is time consuming operation
-
-func init() {
-	resource.AddTestSweepers("equinix_network_ssh_user", &resource.Sweeper{
-		Name: "equinix_network_ssh_user",
-		F:    testSweepNetworkSSHUser,
-	})
-}
 
 type testAccConfig struct {
 	ctx    map[string]interface{}
@@ -39,33 +29,6 @@ func newTestAccConfig(ctx map[string]interface{}) *testAccConfig {
 
 func (t *testAccConfig) build() string {
 	return t.config
-}
-
-func testSweepNetworkSSHUser(region string) error {
-	config, err := sweep.SharedConfigForRegion(region)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting configuration for sweeping Network SSH users: %s", err)
-	}
-	if err := config.Load(context.Background()); err != nil {
-		log.Printf("[INFO][SWEEPER_LOG] error loading configuration: %s", err)
-		return err
-	}
-	users, err := config.Ne.GetSSHUsers()
-	if err != nil {
-		log.Printf("[INFO][SWEEPER_LOG] error fetching NetworkSSHUser list: %s", err)
-		return err
-	}
-	for _, user := range users {
-		if !sweep.IsSweepableTestResource(ne.StringValue(user.Username)) {
-			continue
-		}
-		if err := config.Ne.DeleteSSHUser(ne.StringValue(user.UUID)); err != nil {
-			log.Printf("[INFO][SWEEPER_LOG] error deleting NetworkSSHUser resource %s (%s): %s", ne.StringValue(user.UUID), ne.StringValue(user.Username), err)
-		} else {
-			log.Printf("[INFO][SWEEPER_LOG] sent delete request for NetworkSSHUser resource %s (%s)", ne.StringValue(user.UUID), ne.StringValue(user.Username))
-		}
-	}
-	return nil
 }
 
 func testAccNeSSHUserExists(resourceName string, user *ne.SSHUser) resource.TestCheckFunc {
