+++ internal/sweep/sweep.go
@@ -15,6 +15,7 @@ const (
 	testResourcePrefix        = "tfacc"
 	cannotConvertTimeoutToInt = "cannot convert value of '%s' env variable to int"
 	missingFabricSecrets      = "missing fabric clientId - %s, and clientSecret - %s"
+	missingSecrets            = "missing Equinix credentials: set one of '%s', the pair '%s' and '%s', or '%s' and a subject token in '%s'"
 )
 
 var (
@@ -53,5 +54,43 @@ func GetConfigForFabric() (*config.Config, error) {
 		ClientID:       clientId,
 		ClientSecret:   clientSecret,
 		RequestTimeout: time.Duration(clientTimeoutInt) * time.Second,
+	}, nil
+}
+
+// GetConfig returns a provider configuration for sweepers that call the
+// Equinix API. Credentials are not product-specific: any of an API token, a
+// client ID and secret pair, or an STS token exchange scope and subject token
+// will do.
+func GetConfig() (*config.Config, error) {
+	endpoint := env.GetWithDefault(config.EndpointEnvVar, config.DefaultBaseURL)
+	clientToken := env.GetWithDefault(config.ClientTokenEnvVar, "")
+	clientID := env.GetWithDefault(config.ClientIDEnvVar, "")
+	clientSecret := env.GetWithDefault(config.ClientSecretEnvVar, "")
+	tokenExchangeScope := env.GetWithDefault(config.TokenExchangeScopeEnvVar, "")
+	subjectTokenEnvVar := env.GetWithDefault(config.TokenExchangeSubjectTokenEnvVarEnvVar, config.DefaultTokenExchangeSubjectTokenEnvVar)
+	subjectToken := env.GetWithDefault(subjectTokenEnvVar, "")
+
+	if clientToken == "" && (clientID == "" || clientSecret == "") && (tokenExchangeScope == "" || subjectToken == "") {
+		return nil, fmt.Errorf(missingSecrets,
+			config.ClientTokenEnvVar, config.ClientIDEnvVar, config.ClientSecretEnvVar,
+			config.TokenExchangeScopeEnvVar, subjectTokenEnvVar)
+	}
+
+	clientTimeout := env.GetWithDefault(config.ClientTimeoutEnvVar, strconv.Itoa(config.DefaultTimeout))
+	clientTimeoutInt, err := strconv.Atoi(clientTimeout)
+	if err != nil {
+		return nil, fmt.Errorf(cannotConvertTimeoutToInt, config.ClientTimeoutEnvVar)
+	}
+
+	return &config.Config{
+		BaseURL:                         endpoint,
+		Token:                           clientToken,
+		ClientID:                        clientID,
+		ClientSecret:                    clientSecret,
+		RequestTimeout:                  time.Duration(clientTimeoutInt) * time.Second,
+		TokenExchangeScope:              tokenExchangeScope,
+		TokenExchangeSubjectTokenEnvVar: subjectTokenEnvVar,
+		TokenExchangeSubjectToken:       subjectToken,
+		StsBaseURL:                      env.GetWithDefault(config.StsEndpointEnvVar, config.DefaultStsBaseURL),
 	}, nil
 }
